backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
asyncpg==0.28.0
openai==1.10.0
qdrant-client==1.7.3
python-dotenv==1.0.0